requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
python-dotenv==0.10.3
requests==2.22.0
requests-cache==0.5.2